Analyze_Empathy/requirements.txt
av==10.0.0
blosc2==2.2.9
category-encoders==2.6.3
celluloid==0.2.0
certifi==2023.7.22
charset-normalizer==3.3.0
colorama==0.4.6
contourpy==1.1.1
cycler==0.12.1
dash==2.14.1
deepdish==0.3.7
easing-functions==1.0.4
filelock==3.12.4
fonttools==4.43.1
fsspec==2023.9.2
gaze-tracking==0.0.1
h5py==3.10.0
idna==3.4
imageio==2.31.5
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.5
kornia==0.7.0
lazy_loader==0.3
liwc==0.5.0
lxml==4.9.3
MarkupSafe==2.1.3
matplotlib==3.6.0
mpmath==1.3.0
msgpack==1.0.7
ndindex==1.7
networkx==3.1
nibabel==5.1.0
nilearn==0.10.2
nltools==0.4.7
numexpr==2.8.4
numpy==1.26.1
opencv-python==4.9.0.80
packaging==23.2
pandas==2.1.1
Pillow==10.0.1
plotly-resampler==0.9.1
pmdarima==2.0.4
praat-parselmouth==0.4.3
py-cpuinfo==9.0.0
py-feat==0.6.1
pycaret==3.2.0
pynv==0.3
pyod==1.1.2
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyWavelets==1.4.1
requests==2.31.0
scikit-image==0.22.0
scikit-learn==1.3.1
scikit-plot==0.3.7
scipy==1.11.3
seaborn==0.13.0
six==1.16.0
sktime==0.21.1
sympy==1.12
tables==3.9.1
tbats==1.1.3
threadpoolctl==3.2.0
tifffile==2023.9.26
torch==2.1.0
torchvision==0.16.0
tqdm==4.66.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.6
xgboost==2.0.0
yellowbrick==1.5
